Began implementation of md17/qm9/kaggle models.

+++ src/cormorant/data/collate.py
@@ -78,7 +78,7 @@ def collate_fn(batch):
     atom_mask = batch['charges'] > 0
     edge_mask = atom_mask.unsqueeze(1) * atom_mask.unsqueeze(2)
 
-    batch['atom_mask'] = atom_mask.to(torch.uint8)
-    batch['edge_mask'] = edge_mask.to(torch.uint8)
+    batch['atom_mask'] = atom_mask
+    batch['edge_mask'] = edge_mask
 
     return batch
